+++ lib/src/di/app_dependencies.dart
@@ -2,6 +2,7 @@ import 'package:where_to_go_today/src/core/services/network/dio/dio_module.dart'
 import 'package:where_to_go_today/src/core/ui/errors_handling/scenario_error_handler/scenario_error_handler.dart';
 import 'package:where_to_go_today/src/core/ui/errors_handling/scenario_error_handler/scenarios/snackbar_error_scenarios.dart';
 import 'package:where_to_go_today/src/core/ui/messages/default_message_controller.dart';
+import 'package:where_to_go_today/src/features/auth/services/storage/token_storage.dart';
 import 'package:where_to_go_today/src/features/settings/service/event/settings_event.dart';
 import 'package:where_to_go_today/src/features/settings/service/repository/settings_repository.dart';
 import 'package:where_to_go_today/src/features/settings/service/settings_bloc.dart';
@@ -13,6 +14,7 @@ import 'base/dependency_bundle.dart';
 class AppDependencies extends DependencyBundle {
   final dio = DioModule().dio;
   final settingsController = SettingsBloc(SettingsRepository());
+  final tokenStorage = TokenStorage();
 
   late final messageController = DefaultMessageController();
   late final errorHandler = ScenarioErrorHandler(
@@ -23,5 +25,6 @@ class AppDependencies extends DependencyBundle {
 
   Future<void> init() async {
     settingsController.add(LoadSettings());
+    await tokenStorage.init();
   }
 }